requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
pywatchman==2.0.0
requests==2.32.3
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.3/libexec/wheel-0.43.0-py3-none-any.whl#sha256=80a49545e5e5ef2abeb79ec254447300d41dfb76914fc65efa6da8dd0a766052
